refactor: Splits the app's settings store
